q16

+++ Lesson-22-Activity/Lesson-22-Activity/Main.java
@@ -7,7 +7,7 @@ class Main {
 
   void init(){
 
-    //Using this array to test your summArray and avgArray function
+    //Using this array to test your sumArray and avgArray function
     int nums[] = {10,15,22,54,32,34,65,87,45,100,23,99,95,87,56,83};
     print(sumArray(nums));
     print(avgArray(nums));
@@ -22,7 +22,7 @@ class Main {
   //Problem 1:
   // Write a function called sumArray that accepts an array of integers and
   // returns the sum of all the values in the array.
-  double sumArray(int[] num){
+  int sumArray(int[] num){
     int sum = 0;
     for(int i = 0; i < num.length; i++){
       sum =+ num[i];
